# internal/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Scratchpad Engine API",
    "version": "1.0.0",
    "description": "Multi-platform automation API for Chrome (CDP) and Android (ADB).\n\nEvery error response — over HTTP, WebSocket, or MCP — uses the same typed `ErrorResponse` envelope (fields: `type`, `message`, `code`, `hint`, `request_id`). The `request_id` is generated by the request-id middleware, echoed back in the `X-Request-ID` response header, and repeated in every error envelope so one failure is traceable end to end.\n\n**Honest scope note:** the REST surface is currently a subset of the WebSocket/MCP surface. `POST /api/v1/sessions/{id}/actions` validates and documents six action types (`navigate`, `observe`, `click`, `type`, `scroll`, `wait`). A raw `protocol.ActionRequest` body is also passed through to the engine, but full ActionRequest parity over REST is a later milestone. See the WebSocket endpoints and the `ActionRequest` schema for the full engine action surface."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server (default bind port)"
    }
  ],
...
  "paths": {
...
    "/api/v1/sessions/{id}/actions": {
      "post": {
        "tags": ["Actions"],
        "summary": "Run an action",
        "description": "Executes an action against a session and returns the resulting observation.\n\nThe body may be one of three shapes:\n1. **Typed envelope** — `{\"action\": {\"type\": \"navigate|observe|click|type|scroll|wait\", ...}}` (the validated, documented REST surface).\n2. **Raw InitializeRequest** — `{\"url\": \"https://...\", \"viewport\": {...}}` navigates the session.\n3. **Raw protocol.ActionRequest** — `{\"action\": \"hover\", \"selector\": {...}}` is passed straight through to the engine. This works for the full action surface, but it is undocumented pass-through behavior: parity is a later milestone.\n\nEvery success returns an ObservationResponse (a `delta` when the tree changed by less than a full snapshot).",
        "operationId": "runAction",
        "parameters": [{ "$ref": "#/components/parameters/SessionID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  { "$ref": "#/components/schemas/ActionEnvelope" },
                  { "$ref": "#/components/schemas/InitializeRequest" },
                  { "$ref": "#/components/schemas/ActionRequest" }
                ]
              },
              "examples": {
                "navigate": {
                  "summary": "Open a URL",
                  "value": { "action": { "type": "navigate", "url": "https://example.com" } }
                },
                "observe": {
                  "summary": "Snapshot the current page",
                  "value": { "action": { "type": "observe" } }
                },
                "click": {
                  "summary": "Click at coordinates",
                  "value": { "action": { "type": "click", "x": 640, "y": 480 } }
                },
                "type": {
                  "summary": "Type text",
                  "value": { "action": { "type": "type", "text": "2500*dark" } }
                },
                "wait": {
                  "summary": "Wait for network idle",
                  "value": { "action": { "type": "wait", "timeout_ms": 5000 } }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Observation after the action",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObservationResponse" }
              }
            }
          },
          "400": {
            "description": "Malformed body or unsupported action type",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "request_id": "1a2b3c4d-5e6f-7890-abcd-ef1234567890",
                  "code": "invalid_request",
                  "type": "warning",
                  "message": "unsupported action type: \"screenshot\"",
                  "hint": "Send one of: navigate, observe, click, type, scroll, wait."
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Action or observation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "SessionID": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Session id returned by createSession.",
        "schema": { "type": "string" }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Typed error envelope used by every transport (HTTP, WebSocket, MCP). Carries a stable machine code, a human hint, and the request id for correlation.",
        "properties": {
          "request_id": {
            "type": "string",
            "description": "Correlation id stamped by the request-id middleware. Echo it back via the X-Request-ID header to correlate retries."
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code (e.g. selector_no_match, timeout)."
          },
          "type": {
            "type": "string",
            "enum": ["fatal", "action", "warning"],
            "description": "Severity: fatal = session is broken, action = this action failed but can retry, warning = succeeded with issues."
          },
          "message": { "type": "string" },
          "action": {
            "type": "string",
            "description": "The action that was being attempted (empty for session-level errors)."
          },
          "selector": {
            "$ref": "#/components/schemas/Selector",
            "description": "The selector that failed, when applicable."
          },
          "screenshot": {
            "type": "string",
            "format": "byte",
            "description": "Base64 JPEG captured at the moment of failure (WebSocket path)."
          },
          "hint": {
            "type": "string",
            "description": "What the agent should try next."
          }
        },
        "required": ["type", "message"]
      },
      "Selector": {
        "type": "object",
        "description": "Structured locator. Resolution order: CSS > XPath > text > role > test_id > placeholder. Pass multiple strategies and the best match wins.",
        "properties": {
          "css": { "type": "string", "example": "#submit" },
          "xpath": { "type": "string", "example": "//button[text()='Submit']" },
          "text": { "type": "string", "example": "Sign in" },
          "role": { "type": "string", "example": "button" },
          "test_id": { "type": "string", "example": "login-submit" },
          "placeholder": { "type": "string", "example": "Email" }
        }
      },
      "Viewport": {
        "type": "object",
        "properties": {
          "width": { "type": "integer", "example": 1280 },
          "height": { "type": "integer", "example": 720 }
        },
        "required": ["width", "height"]
      },
      "InitializeRequest": {
        "type": "object",
        "description": "Navigates the session to a URL. Also accepted directly as the body of POST /api/v1/sessions/{id}/actions.",
        "properties": {
          "url": { "type": "string", "example": "https://example.com" },
          "viewport": { "$ref": "#/components/schemas/Viewport" }
        },
        "required": ["url"]
      },
      "ActionEnvelope": {
        "type": "object",
        "description": "The typed action envelope accepted by the REST actions endpoint.",
        "properties": {
          "action": { "$ref": "#/components/schemas/ActionPayload" }
        },
        "required": ["action"]
      },
      "ActionPayload": {
        "type": "object",
        "description": "The typed action payload validated by the REST API. Only six actions are supported today: navigate, observe, click, type, scroll, wait. Full ActionRequest parity over REST is a later milestone.",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["navigate", "observe", "click", "type", "scroll", "wait"]
          },
          "url": { "type": "string", "description": "navigate: URL to open (http://... or an Android package name)." },
          "x": { "type": "integer", "description": "click/scroll: X coordinate." },
          "y": { "type": "integer", "description": "click/scroll: Y coordinate." },
          "text": { "type": "string", "description": "type: text to enter." },
          "delta_x": { "type": "integer", "description": "scroll: horizontal scroll amount." },
          "delta_y": { "type": "integer", "description": "scroll: vertical scroll amount." },
          "timeout_ms": { "type": "integer", "description": "click/scroll/wait: timeout in milliseconds." }
        },
        "required": ["type"]
      },
      "ActionRequest": {
        "type": "object",
        "description": "Full protocol ActionRequest (the WebSocket/MCP action surface). Over REST it is accepted in raw form and passed through to the engine — the full surface works, but only the six actions in ActionPayload are validated and documented today; parity is a later milestone.",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "click", "type", "scroll", "wait", "switch_tab", "close_tab", "check", "uncheck",
              "submit_form", "fill_form", "dismiss_modal", "hover", "double_click", "right_click",
              "drag_drop", "select_option", "press_key_combo", "execute_js", "scroll_into_view",
              "switch_to_iframe", "accept_dialog", "dismiss_dialog", "upload_file", "set_geolocation",
              "mock_network_response", "assert"
            ]
          },
          "target_id": { "type": "string", "description": "Tab/target id for tab actions." },
          "x": { "type": "integer", "description": "Coordinate-based actions: X." },
          "y": { "type": "integer", "description": "Coordinate-based actions: Y." },
          "text": { "type": "string", "description": "type/assert: text." },
          "delta_x": { "type": "integer", "description": "scroll: horizontal delta." },
          "delta_y": { "type": "integer", "description": "scroll: vertical delta." },
          "condition": { "type": "string", "description": "wait: e.g. network_idle, url_match." },
          "timeout_ms": { "type": "integer", "description": "Auto-wait/assert timeout in milliseconds (default 10000)." },
          "selector": { "$ref": "#/components/schemas/Selector" },
          "target_selector": { "$ref": "#/components/schemas/Selector", "description": "drag_drop: destination selector." },
          "pattern": { "type": "string", "description": "Regex/text pattern for waits and assertions." },
          "attribute": { "type": "string", "description": "Attribute name for attribute assertions." },
          "value": { "type": "string", "description": "Expected value for value/attribute assertions." },
          "option_value": { "type": "string", "description": "select_option: option value." },
          "option_text": { "type": "string", "description": "select_option: option label." },
          "js": { "type": "string", "description": "execute_js: JavaScript to run." },
          "dialog_action": { "type": "string", "description": "accept/dismiss_dialog action." },
          "upload_files": { "type": "array", "items": { "$ref": "#/components/schemas/UploadFile" } },
          "key_chord": { "$ref": "#/components/schemas/KeyChord" },
          "geolocation": { "$ref": "#/components/schemas/Geolocation" },
          "network_mock": { "$ref": "#/components/schemas/NetworkMock" },
          "iframe_selector": { "$ref": "#/components/schemas/Selector", "description": "switch_to_iframe: iframe locator." },
          "assertion": { "$ref": "#/components/schemas/AssertionRequest" },
          "tab_id": { "type": "string", "description": "switch_tab / close_tab: target tab id." },
          "form_fields": { "type": "array", "items": { "$ref": "#/components/schemas/FormField" } },
          "modal_strategy": {
            "type": "string",
            "enum": ["auto", "click_outside", "press_escape", "click_button"],
            "description": "dismiss_modal strategy."
          }
        },
        "required": ["action"]
      },
      "AssertionRequest": {
        "type": "object",
        "description": "Web-first assertion: the engine polls until the condition holds or the retry timeout elapses.",
        "properties": {
          "type": {
            "type": "string",
            "description": "element_exists, element_visible, element_checked, element_count, text_equals, text_contains, text_matches, attr_equals, attr_contains, attr_matches, value_equals, page_title, page_url, url_matches, console_error_count, network_request_status, network_no_errors, screenshot_matches."
          },
          "selector": { "$ref": "#/components/schemas/Selector" },
          "text": { "type": "string" },
          "attribute": { "type": "string" },
          "value": { "type": "string" },
          "checked": { "type": "boolean", "description": "Expected checkbox/radio state." },
          "pattern": { "type": "string", "description": "Regex/text pattern." },
          "expected_count": { "type": "integer" },
          "timeout_ms": { "type": "integer", "description": "Retry window in ms (default 5000)." },
          "tolerance": { "type": "integer", "description": "Perceptual-diff tolerance." },
          "screenshot_base64": { "type": "string", "description": "Baseline image for screenshot_matches." },
          "screenshot_tolerance": { "type": "integer" }
        },
        "required": ["type"]
      },
      "UploadFile": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "content_base64": { "type": "string" }
        }
      },
      "KeyChord": {
        "type": "object",
        "properties": {
          "key": { "type": "string", "example": "a" },
          "ctrl": { "type": "boolean" },
          "alt": { "type": "boolean" },
          "shift": { "type": "boolean" },
          "meta": { "type": "boolean" }
        }
      },
      "Geolocation": {
        "type": "object",
        "properties": {
          "latitude": { "type": "number", "format": "double", "example": 37.7749 },
          "longitude": { "type": "number", "format": "double", "example": -122.4194 },
          "accuracy_m": { "type": "number", "format": "double" }
        },
        "required": ["latitude", "longitude"]
      },
      "NetworkMock": {
        "type": "object",
        "properties": {
          "url_pattern": { "type": "string" },
          "method": { "type": "string" },
          "status": { "type": "integer" },
          "headers": { "type": "object", "additionalProperties": { "type": "string" } },
          "body_base64": { "type": "string" }
        }
      },
      "FormField": {
        "type": "object",
        "properties": {
          "selector": { "$ref": "#/components/schemas/Selector" },
          "value": { "type": "string" }
        },
        "required": ["selector", "value"]
      },
      "ObservationResponse": {
        "type": "object",
        "description": "Snapshot of the current page/screen returned after every action and observation. When the tree changed by less than a full snapshot, `type` is \"delta\" and `delta` carries the change instead of a full `spatial_tree`.",
        "properties": {
          "type": { "type": "string", "enum": ["observation", "delta"] },
          "system_state": { "$ref": "#/components/schemas/SystemState" },
          "viewport": { "$ref": "#/components/schemas/Viewport" },
          "visual_context": { "type": "string", "description": "Base64-encoded screenshot." },
          "spatial_tree": { "type": "array", "items": { "$ref": "#/components/schemas/SpatialNode" } },
          "delta": { "$ref": "#/components/schemas/TreeDelta" },
          "logs": { "type": "array", "items": { "$ref": "#/components/schemas/ConsoleLog" } },
          "page_info": { "$ref": "#/components/schemas/PageInfo" },
          "tabs": { "type": "array", "items": { "$ref": "#/components/schemas/TabInfo" } },
          "action_result": { "$ref": "#/components/schemas/ActionResult" },
          "assertion_result": { "$ref": "#/components/schemas/AssertionResult" }
        }
      },
      "SystemState": {
        "type": "object",
        "properties": {
          "document_status": { "type": "string", "example": "complete" },
          "inflight_requests": { "type": "integer", "example": 0 }
        }
      },
      "SpatialNode": {
        "type": "object",
        "description": "A UI element in the accessibility tree.",
        "properties": {
          "node_id": { "type": "string" },
          "role": { "type": "string", "description": "ARIA role (Chrome) or class name (Android)." },
          "name": { "type": "string", "description": "Text or accessible name." },
          "bounds": { "$ref": "#/components/schemas/Bounds" },
          "scroll_state": { "$ref": "#/components/schemas/ScrollState" },
          "children": { "type": "array", "items": { "$ref": "#/components/schemas/SpatialNode" } },
          "interactive": { "type": "boolean", "description": "True for actionable elements (buttons, links, inputs)." },
          "value": { "type": "string", "description": "Current value of a form control." },
          "description": { "type": "string", "description": "aria-description or title attribute." }
        }
      },
      "Bounds": {
        "type": "object",
        "properties": {
          "x": { "type": "number", "format": "double" },
          "y": { "type": "number", "format": "double" },
          "width": { "type": "number", "format": "double" },
          "height": { "type": "number", "format": "double" }
        }
      },
      "ScrollState": {
        "type": "object",
        "properties": {
          "can_scroll_down": { "type": "boolean" },
          "can_scroll_up": { "type": "boolean" },
          "current_percentage": { "type": "integer" }
        }
      },
      "TreeDelta": {
        "type": "object",
        "properties": {
          "added": { "type": "array", "items": { "$ref": "#/components/schemas/SpatialNode" } },
          "removed": { "type": "array", "items": { "type": "string" }, "description": "node_ids removed." },
          "updated": { "type": "array", "items": { "$ref": "#/components/schemas/SpatialNode" } }
        }
      },
      "ConsoleLog": {
        "type": "object",
        "properties": {
          "level": { "type": "string", "example": "info" },
          "message": { "type": "string" },
          "timestamp": { "type": "integer", "description": "Unix milliseconds." }
        }
      },
...
      "PageInfo": {
        "type": "object",
        "description": "Current page/screen metadata.",
        "properties": {
          "url": { "type": "string" },
          "title": { "type": "string" },
          "platform": { "type": "string", "enum": ["web", "android"] },
          "load_status": { "type": "string" },
          "navigation_id": { "type": "integer" },
          "dialog_state": { "type": "string" },
          "tab_count": { "type": "integer" },
          "extra": { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
      "TabInfo": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string" },
          "title": { "type": "string" },
          "active": { "type": "boolean" },
          "opener_id": { "type": "string" }
        }
      },
      "ActionResult": {
        "type": "object",
        "description": "Rich feedback returned after every action.",
        "properties": {
          "success": { "type": "boolean" },
          "action": { "type": "string" },
          "error": { "type": "string" },
          "elapsed_ms": { "type": "integer" },
          "action_metadata": { "type": "object", "additionalProperties": {} },
          "screenshot": { "type": "string", "description": "Base64 JPEG captured after the action." },
          "element_highlight": { "type": "string", "description": "Base64 JPEG with the targeted element outlined." }
        }
      },
      "AssertionResult": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "type": { "type": "string" },
          "message": { "type": "string" },
          "elapsed_ms": { "type": "integer" },
          "attempts": { "type": "integer", "description": "How many times the assertion was polled before resolving." },
          "poll_interval_ms": { "type": "integer" },
          "extra": { "type": "object", "additionalProperties": {} }
        }
      },
...
    }
  }
}